refactor: move cli args to separate crates
