+++ bin/katana/Cargo.toml
@@ -7,30 +7,24 @@ repository.workspace = true
 version.workspace = true
 
 [dependencies]
-katana-core.workspace = true
 katana-db.workspace = true
 katana-node.workspace = true
 katana-primitives.workspace = true
-katana-slot-controller = { workspace = true, optional = true }
+katana-cli.workspace = true
 
-alloy-primitives.workspace = true
 anyhow.workspace = true
 byte-unit = "5.1.4"
-cainome-cairo-serde.workspace = true
 clap.workspace = true
 clap_complete.workspace = true
 comfy-table = "7.1.1"
 console.workspace = true
 dojo-utils.workspace = true
-serde.workspace = true
 serde_json.workspace = true
 shellexpand = "3.1.0"
 tokio.workspace = true
-toml.workspace = true
 tracing.workspace = true
 tracing-log.workspace = true
 tracing-subscriber.workspace = true
-url.workspace = true
 
 [dev-dependencies]
 assert_matches.workspace = true
@@ -40,5 +34,5 @@ starknet.workspace = true
 default = [ "jemalloc", "slot" ]
 
 jemalloc = [  ]
-slot = [ "dep:katana-slot-controller", "katana-primitives/slot" ]
+slot = [ "katana-primitives/slot", "katana-cli/slot" ]
 starknet-messaging = [ "katana-node/starknet-messaging" ]
