+++ bin/katana/src/cli/mod.rs
@@ -1,10 +1,10 @@
 mod db;
 mod node;
-mod options;
 
 use anyhow::Result;
 use clap::{Args, CommandFactory, Parser, Subcommand};
 use clap_complete::Shell;
+use katana_cli::NodeArgs;
 use katana_node::version::VERSION;
 
 #[derive(Parser)]
@@ -14,7 +14,7 @@ pub struct Cli {
     commands: Option<Commands>,
 
     #[command(flatten)]
-    node: node::NodeArgs,
+    node: NodeArgs,
 }
 
 impl Cli {
@@ -26,7 +26,7 @@ impl Cli {
             };
         }
 
-        self.node.with_config_file()?.execute()
+        node::execute(&self.node.with_config_file()?)
     }
 }
