+++ bin/katana/src/main.rs
@@ -1,7 +1,6 @@
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
 mod cli;
-mod utils;
 
 use anyhow::Result;
 use clap::Parser;
