+++ crates/katana/cli/src/lib.rs
@@ -1,0 +1,8 @@
+#![cfg_attr(not(test), warn(unused_crate_dependencies))]
+
+pub mod node;
+pub mod options;
+pub mod utils;
+
+pub use node::{NodeArgs, NodeArgsConfig};
+pub use options::*;
