+++ crates/katana/cli/test-data/genesis.json
