+++ crates/torii/cli/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+edition.workspace = true
+license.workspace = true
+name = "torii-cli"
+repository.workspace = true
+version.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+clap.workspace = true
+dojo-utils.workspace = true
+serde.workspace = true
+starknet.workspace = true
+torii-core.workspace = true
+toml.workspace = true
+url.workspace = true
+
+[dev-dependencies]
+assert_matches.workspace = true
+camino.workspace = true
