+++ crates/torii/cli/src/lib.rs
@@ -1,0 +1,7 @@
+#![cfg_attr(not(test), warn(unused_crate_dependencies))]
+
+pub mod args;
+pub mod options;
+
+pub use args::ToriiArgs;
+pub use options::*;
